Replace BinaryOperation
