+++ core/src/main/scala/latis/ops/BinaryOperation.scala
@@ -1,8 +1,16 @@
 package latis.ops
 
-import latis.data.Data
+import cats.effect.IO
+import fs2.Stream
+
+import latis.data.Sample
 import latis.model.DataType
 import latis.util.LatisException
+
+
+//TODO: replace BinaryOperation
+//  this adds the models to applyToData
+//  also experimenting with using a Stream instead of Data
 
 /**
  * Defines an Operation that combines two Datasets into one.
@@ -14,16 +22,21 @@ trait BinaryOperation extends Operation {
    Then we could drop "Unary"
    Would this give up opportunities to fold a collection of datasets?
    */
+  
+  //TODO: combine(ds1, ds2)?
 
-  /**
-   * Combines the models of two Datasets.
-   */
-  def applyToModel(model1: DataType, model2: DataType): Either[LatisException, DataType]
+  /** Combines the models of two Datasets. */
+  def applyToModel(
+    model1: DataType, 
+    model2: DataType
+  ): Either[LatisException, DataType]
 
-  /**
-   * Combines the Data of two Datasets.
-   */
-  //TODO: need model
-  def applyToData(data1: Data, data2: Data): Either[LatisException, Data]
+  /** Combines the Data of two Datasets. */
+  def applyToData(
+    model1: DataType,
+    stream1: Stream[IO, Sample],
+    model2: DataType,
+    stream2: Stream[IO, Sample]
+  ): Either[LatisException, Stream[IO, Sample]]
 
 }
